repo/backend/requirements.txt
# Django core
Django==5.1.4
djangorestframework==3.15.2
django-celery-beat==2.7.0

# Database
mysqlclient==2.2.6

# Auth & Security
argon2-cffi==23.1.0
cryptography==44.0.0
python-magic==0.4.27

# Async tasks
celery[redis]==5.4.0
redis==5.2.1

# File processing
Pillow==11.1.0
reportlab==4.2.5
openpyxl==3.1.5

# Logging
structlog==24.4.0
python-json-logger==2.0.7

# Server
gunicorn==23.0.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
